Upgrade EML to SOSO v1.3.2 conversion
Update the conversion of EML metadata to the latest SOSO standard
(v1.3.2). Utilize the `soso` Python package for the conversion.

Note, the `soso` package is currently only available on the development
branch of the GitHub repository.

Drop the use of the `includedInDataCatalog` property when referring to
the entire repository holdings. This property is intended to reference a
subset of repository holdings belonging to a thematic subset. See SOSO
conventions for details.

Simplify `publisher` and `provider` properties according to conventions,
allowing the use of the repository identifier, which cross-references the
repository metadata defined in the repository markup.

Closes #7