Update environment.yml and requirements.txt via export
This commit updates `environment.yml` and `requirements.txt` by exporting
the built environment from `environment-min.yml`. This reflects the
preferred development workflow of pinning packages in
`environment-min.yml` and then generating the other files.

This update was triggered by the need to pin the `soso` package
(related to #33) and indirectly addresses Dependabot-raised
CVEs for `setuptools` and `flask` packages.